// src/api/types/IngestRemoteDocument.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as GroundX from "../index";

export interface IngestRemoteDocument {
    /** The bucketId of the bucket which this remote file will be ingested into. */
    bucketId: number;
    /** The name of the file being ingested. */
    fileName?: string;
    fileType?: GroundX.DocumentType;
    /** A dictionary of key-value pairs that can be used to pre-filter documents prior to a search. */
    filter?: Record<string, unknown>;
    processLevel?: GroundX.ProcessLevel;
    /** Custom metadata which can be used to influence GroundX's search functionality. This data can be used to further hone GroundX search. */
    searchData?: Record<string, unknown>;
    /** The URL of the document being ingested by GroundX. */
    sourceUrl: string;
}